feat: add update command to CLI
This commit introduces a new 'update' command to the Hyphen CLI. This command enables users to update the CLI to a target version or the latest version available for their specific operating system. Internally, it fetches the update file from a specified URL, validates the operating system, and performs the update. It prints out a success message upon successful update or a failure reason if the update didn't go as expected.

This change impacts the root.go and introduces a new update.go file under the cmd/update/ directory.

NOTE: The update URL utilized in this commit directs to 'localhost:4000'. Future versions should direct this to the production API.