feat(update): use temp file and better handling for windows in cmd/update/update.go
This commit modifies the  function in cmd/update/update.go to use a temporary file for downloading, rather than directly saving to the target filename. This reduces the chance of ending up with a partial or corrupt download if the process is interrupted. The permission of the downloaded file is also set before renaming it to the target filename.

For windows, we now use a batch script to replace the file after the process exits to ensure the program is not in use when we try to update it.

This commit also adds two helper functions getExecutablePath and replaceFile. The first gets the path of the currently running executable, while the second is a simple wrap around os.Rename to semantically differentiate renaming the downloaded file to its final location from other renames that may occur in the code.