feat: move gitignore check and update to initialization
Changes were made in several files for this commit. In 'cmd/env/create/create.go', the check and possible update of the .gitignore file was removed from the create function. This has been moved to the 'cmd/env/initialize/initialize.go' file to occur during initialization instead. Additionally, in 'internal/environment/environment.go', the behavior of EnsureGitignore was changed to create a .gitignore file and add an entry if one does not already exist. If the .git directory does not exist, the function will exit early.