feat(create.go): change argument requirement in CreateCmd
Changed the Arguments requirement from ExactArgs(1) to MaximumNArgs(1) in CreateCmd within cmd/env/create/create.go. This allows for the environment argument to be optional. Initialized env with an empty string and added a condition to set env value if it's provided in arguments. Minor code reformat and comment updates included.