feat: add tag-based trigger and upload to R2 in CI workflow
This commit includes changes in .github/workflows/ci.yml file to trigger the CI workflow on push event for tags that match a specific pattern, in addition to pushes on the main branch. An 'upload' job is added to this CI workflow that is dependent on the 'build' job and runs only when a tag is pushed. This 'upload' job uses the new workflow added in the .github/workflows/upload-to-r2.yml file.

The 'upload-to-r2.yml' is a new file that defines a workflow for uploading the built artifacts to R2 cloud storage. This workflow runs on a workflow call event, takes the 'git_tag' as an input and uses it while uploading the artifact. AWS CLI is used to interact with the R2 cloud storage and the necessary configurations for AWS CLI are set during the run.