feat: add package name environment variable for dynamic file uploads
We've added a new environment variable PACKAGE-NAME in the GitHub Action workflow upload-to-r2-dev.yml file. By doing so, we can now dynamically change the package name for uploading files to the S3 storage bucket. This will streamline and give flexibility to the deployment process. Updates have been also made to verify the uploaded file names reflecting the new change.