feat: restructure CI/CD workflow and add R2 upload
Refactor the Continuous Integration and Continuous Deployment (CI/CD) workflow by adding a dedicated workflow for deployments to both the Dev and Production environments. Remove build and upload-dev jobs from the ci.yml workflow file and move them to new dedicated files deployment-to-dev.yml and deployment-to-prod.yml. Include a new upload-to-r2.yml workflow file in the workflow to handle the upload of artifacts to R2.