feat: improve input validation in env initialize command
The updates in this commit add additional validation to the PromptForAppName and PromptForAppId functions in the initialize.go file. An empty app name is now not allowed and the user will be prompted to enter a valid app name. Also, the code is updated to give an alert if the user's response is neither 'yes' nor 'no' when asked to overwrite.