feat: Added function to create alias in install.sh
In this commit a function was added to the install.sh file that creates an alias for the hyphen command. The function will attempt to add the alias to either .zshrc or .bashrc files (if they exists) in the user's home directory. If neither file exists, it tries to add the alias to .bash_profile. If it fails to find any of these files in the user's home directory, it advises the user to manually add the alias to their shell configuration file. Following this, please restart your terminal or source the relevant configuration file(s) to enable changes.