feat: Add GitHub workflows for publishing to Chocolatey and uploading install script to R2
Add two new workflows to automate package publishing and installation process. The first pipeline publish-to-chocolatey.yml is for publishing the package to Chocolatey whenever a new version is released. The second pipeline upload-install-script-to-r2.yml is for uploading the installation script to R2 whenever there is a change to 'install.sh' on the main branch.

Also, added new files hyphen.nuspec to defines the package metadata for Chocolatey package and chocolateyinstall.ps1 to handle chocolatey installation process.