feat: Allow OAuth tokens and expiry times to be stored in manifest
Modified the program to allow OAuth tokens and expiry times to be stored in the app's manifest file. This enables the OAuth service to fetch valid tokens from the manifest and refresh the tokens when they have expired.

Refactored manifest.go and its test to use a mocked filesystem for better test isolation. Replaced all os package file operations with the corresponding functions from the fsutil.FileSystem interface.

This commit also includes a refactor of the way the global config directory is retrieved, which is now OS-agnostic and uses native functions provided by the os package.