feat: add global flag for setting organization and project IDs (#87)
This commit introduced a global flag for 'set-org' and 'set-project' commands, allowing users to set the organization and project ID globally. In addition to the changes in the command files, corresponding new functions were added to the manifest go file to handle the global setting of IDs. Now when updating the organization or project, an additional status message indicates whether the operation was set to global.