feat: Add versioning support to environment variables
Improves the handling of environment variables by adding versioning support. This was achieved by making modifications to the 'push' operation and updating the logic for handling remote and local environment variables. This versioning support allows the program to detect and warn if the local version of an environment is older than the remote version. It also saves the updated version locally after a successful push operation. This change was done across multiple files.

In cmd/env/push/push.go, we corrected a spelling mistake and added new lines for fetching key from manifest, fetching remoteEnv, localEnv incrementation and saving updated version.

In internal/env/env.go, versioning related code was added to handle the parsing of versions from environment file data.

In internal/env/service.go, a modification was made to unset the version before pushing the environment.

This commit is backward compatible, thus versioning is optional. If no version is found in an environment file, the program defaults to version 1. This commit also introduces a new metadata section, within environment files, where the version number is stored.