feat(run.go): refactor logic for handling environment variables
Refactored the logic for handling environment variables in run.go. Created a new error handler for the case case where no command is specified. Included an option to use a default environment if an environment isn't specified. Simplified the process of merging and appending environment variables by removing the use of the mergo package and instead using buffered scanning of the environment data. Errors are now handled locally within function calls instead of being returned up the chain.