feat: added ID field and newManifestSecret function to ManifestSecret… (#125)
* feat: added ID field and newManifestSecret function to ManifestSecret struct

In the file internal/manifest/manifest.go, a new ID field has been added to the ManifestSecret struct and been set to current Unix timestamp using time.Now().Unix(). A new function newManifestSecret has also been created for creating a new ManifestSecret instance. The function takes a pointer to a SecretKey instance as arguments and returns a ManifestSecret instance. The Initialize function has been updated to use newManifestSecret when creating a ManifestSecret.

* feat(env): add rotate key command and refactor pull command

- In env.go, added a new command for key rotation.
- Refactored the pull command in pull.go to improve code clarity and maintainability.
- Created a new file rotatekey.go to handle the key rotation logic.
- Added a SecretKeyId field in the Env structure in env.go to keep track of the used secret key.
- In manifest.go, updated the ManifestSecret structure to include a SecretKeyId and refactored the newManifestSecret function accordingly.
- All these changes are backward compatible and do not break any existing functionalities.

* feat: add user prompt and validation to key rotation process

This commit introduces a new feature to the key rotation process in the 'rotatekey.go' file. It now displays a warning message to the user and prompts them for confirmation before proceeding with the key rotation. If the user confirms, the process continues. Otherwise, it is cancelled. Additionally, error messaging has been improved when restoring the manifest, and success messaging has been added upon completion of the key rotation.

* feat: Added secretKeyId to get and put environment methods

This commit includes changes to how environment variables are retrieved and updated. Specifically, a secretKeyId has been added to the methods for getting and putting environment data. This allows for better encryption and key rotation. Various functions and methods across different files have been updated to accommodate this new parameter. In some cases, methods have been renamed and refactored for better clarity and consistency.

This commit effects the following files:
- cmd/env/pull/pull.go
- cmd/env/push/push.go
- cmd/env/rotatekey/rotatekey.go
- cmd/initialize/initialize.go
- internal/database/database.go
- internal/env/service.go
- internal/env/service_test.go
- internal/manifest/manifest.go

* feat(env): add silent option to various environment commands

Add an option 'Silent' to suppress the output of several environment commands in the files 'pull.go', 'push.go', and 'rotatekey.go'. Modify 'service.go' to correctly generate URL by encoding URL query parameters instead of manually concatenating strings. This change improves code readability and maintainability.

* feat: change warning messages to info in rotatekey.go

This commit updates the warning messages displayed during the encryption key rotation process in the rotatekey.go file to information messages. The messages now display as informational instead of warnings, but the content remains the same.