feat(pull): added versioning support to pull command
This update introduces the ability to specify a version when executing the pull command. It also includes a warning message in case a requested version is not found for an environment. In addition, it modifies error messages for better clarity regarding resource non-existence issues. The changes have been appropriately tested.