feat: add directory validation and update error messages
This commit introduces directory validation to cmd/initialize/initialize.go, and makes updates to the error message displayed in internal/manifest/manifest.go. Specifically, it prevents the initialization of the application in the user's home directory, guides the user to project directory initialization, and modifies the error message when no .hxkey is found.