feat: add check for organization ID consistency in pkg/flags/organization.go
The function 'GetOrganizationID' has been modified to check the consistency between the local and global organization IDs. The function 'isGlobalOrgIdSameAsLocal' has been added to verify if the local and global organization IDs match. Warning is displayed if there is a mismatch. Also, the reference of manifest data variable has been renamed from 'manifest' to 'm'.