Handle TUI in non-Terminal use (e.g. CI) (#228)
* detect if in terminal and only use TUI if so

* fix casing on run message data

* don't just check status

This code was causing the pipeline to not display sometimes because the status was actually running, not pending, by the time the message came in.

* add verbose and error messages to the TUI

* make an initial run call

This initial run call handles the case that the pipeline is already running before we establish the WebSocket connection, and therefore would otherwise mis the run message.

* play nice as a websocket client

* need a line

* check run status on error

Frequently, the unexpected EOF happens after the run has actually succeeded (or failed or been cancelled), so rather than end with an error, we can gracefully exist if the run is actually complete

* also be a good websocket client in non-TUI path

* add the API check on error for non-TUI path

* handle non-TUI in build with Dockerfile

* de-dupe things in deploy for non-TUI

* handle deadlock